# docs/api/topic-e-plan-2-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinBrain OS Plan 2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/team/members": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Team Members",
        "operationId": "team_members_team_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/team/invitations": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Invite Member",
        "operationId": "invite_member_team_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/team/members/{user_id}": {
      "patch": {
        "tags": [
          "team"
        ],
        "summary": "Update Member",
        "operationId": "update_member_team_members__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/members/{user_id}/sign-out": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Sign Out Member",
        "operationId": "sign_out_member_team_members__user_id__sign_out_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignOutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/trust/posture": {
      "get": {
        "tags": [
          "trust"
        ],
        "summary": "Posture",
        "operationId": "posture_trust_posture_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostureResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/trust/guardrail-events": {
      "get": {
        "tags": [
          "trust"
        ],
        "summary": "Guardrail Events",
        "operationId": "guardrail_events_trust_guardrail_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardrailEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataMode": {
        "type": "string",
        "enum": [
          "stub",
          "live"
        ],
        "title": "DataMode",
        "description": "Every contract response says whether its data is a canned stub or computed live."
      },
...
      "GuardrailEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "owasp_code": {
            "$ref": "#/components/schemas/OwaspAgenticRisk"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "blocked",
              "quarantined",
              "escalated"
            ],
            "title": "Outcome"
          }
        },
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "agent_id",
          "owasp_code",
          "title",
          "detail",
          "outcome"
        ],
        "title": "GuardrailEvent"
      },
      "GuardrailEventsResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/GuardrailEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "events"
        ],
        "title": "GuardrailEventsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvitationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "job_functions": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "maxItems": 11,
            "minItems": 1,
            "title": "Job Functions"
          }
        },
        "type": "object",
        "required": [
          "email",
          "role",
          "job_functions"
        ],
        "title": "InvitationRequest"
      },
      "JobFunction": {
        "type": "string",
        "enum": [
          "owner",
          "operations",
          "finance",
          "sales",
          "customer_service",
          "marketing",
          "procurement",
          "logistics",
          "production",
          "hr",
          "compliance"
        ],
        "title": "JobFunction",
        "description": "The positions a person can hold. One person may hold several."
      },
...
      "MemberResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "member": {
            "$ref": "#/components/schemas/TeamMember"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "member"
        ],
        "title": "MemberResponse"
      },
      "MemberUpdateRequest": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_functions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobFunction"
                },
                "type": "array",
                "maxItems": 11,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Functions"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "MemberUpdateRequest"
      },
...
      "OwaspAgenticRisk": {
        "type": "string",
        "enum": [
          "ASI01",
          "ASI02",
          "ASI03",
          "ASI04",
          "ASI05",
          "ASI06",
          "ASI07",
          "ASI08",
          "ASI09",
          "ASI10"
        ],
        "title": "OwaspAgenticRisk",
        "description": "OWASP Top 10 for Agentic Applications (2026)."
      },
...
      "PostureMetric": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "status": {
            "type": "string",
            "enum": [
              "good",
              "attention",
              "risk"
            ],
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "value",
          "status",
          "detail"
        ],
        "title": "PostureMetric"
      },
      "PostureResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/PostureMetric"
            },
            "type": "array",
            "title": "Metrics"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/GuardrailEvent"
            },
            "type": "array",
            "title": "Recent Events"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "score",
          "metrics",
          "recent_events"
        ],
        "title": "PostureResponse"
      },
...
      "SignOutResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "sessions_revoked": {
            "type": "integer",
            "title": "Sessions Revoked"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "user_id",
          "sessions_revoked"
        ],
        "title": "SignOutResponse"
      },
...
      "TeamMember": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "email_masked": {
            "type": "string",
            "title": "Email Masked"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "job_functions": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "title": "Job Functions"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "mfa_enrolled": {
            "type": "boolean",
            "title": "Mfa Enrolled"
          },
          "last_active_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Active At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "display_name",
          "email_masked",
          "role",
          "job_functions",
          "active",
          "mfa_enrolled",
          "last_active_at"
        ],
        "title": "TeamMember"
      },
      "TeamResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/TeamMember"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "members"
        ],
        "title": "TeamResponse"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "general_employee",
          "finance_ops",
          "owner_director",
          "compliance"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}